NN-4835 refactoring hearing outcomes for referral service
